requirements.txt
clip
hydra-core
wandb
datasets
lightning
albumentations
scikit-learn
opencv-python
git+https://github.com/rwightman/pytorch-image-models.git@v0.9.12
git+https://github.com/openai/CLIP.git

# GradCAM visualization dependencies
matplotlib>=3.5.0
pandas>=1.3.0
tqdm>=4.62.0
Pillow>=8.3.0
